# openapi.yaml
openapi: 3.0.3
info:
  title: DT-RAG Hybrid Search API
  version: 1.8.1
  description: |
    Dynamic Taxonomy RAG (DT-RAG) Hybrid Search API combining BM25 keyword search
    with vector similarity search using PostgreSQL Full-text Search and pgvector.

    **Key Features:**
    - Hybrid BM25 + vector search with cross-encoder reranking
    - Taxonomy-based content filtering
    - Real-time performance metrics and analytics
    - Configurable search weights and normalization
    - Search result caching and optimization

    **Performance Targets:**
    - Recall@10 ≥ 0.85
    - Search latency p95 ≤ 1s
    - Cost ≤ ₩3/search

  contact:
    name: DT-RAG API Support
    email: support@dt-rag.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: https://api.dt-rag.example.com/v1
    description: Production server
...
paths:
...
  /ingestion/upload:
    post:
      tags:
        - Ingestion
      summary: Upload document for processing
      description: |
        Upload a document file for ingestion, processing, and indexing.
        Supported formats: PDF, DOCX, CSV, HTML, TXT.

        **Processing Pipeline:**
        1. File validation and format detection
        2. Content extraction with parser
        3. Intelligent chunking (500 tokens, 128 overlap)
        4. PII detection and masking
        5. Embedding generation (OpenAI 1536-dim)
        6. Vector storage with BM25 index

      operationId: uploadDocument
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/DocumentUploadRequest'
            examples:
              pdf_upload:
                summary: Upload PDF document
                value:
                  file: "(binary PDF content)"
                  taxonomy_path: ["Technology", "AI", "Machine Learning"]
                  source_url: "https://example.com/ml-paper.pdf"
                  author: "John Doe"
                  language: "en"
      responses:
        '202':
          description: Document accepted for processing
          headers:
            X-Job-ID:
              description: Unique job identifier for tracking
              schema:
                type: string
                format: uuid
                example: "550e8400-e29b-41d4-a716-446655440005"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionJobResponse'
              example:
                job_id: "550e8400-e29b-41d4-a716-446655440005"
                status: "pending"
                estimated_completion_minutes: 2
                message: "Document accepted for processing"
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalid_format:
                  summary: Unsupported file format
                  value:
                    detail: "Unsupported file format. Supported: pdf, docx, csv, html, txt"
                    correlationId: "550e8400-e29b-41d4-a716-446655440006"
                    timestamp: "2025-01-14T12:00:00Z"
                file_too_large:
                  summary: File size exceeds limit
                  value:
                    detail: "File size exceeds 50MB limit"
                    correlationId: "550e8400-e29b-41d4-a716-446655440007"
                    timestamp: "2025-01-14T12:00:00Z"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /ingestion/status/{jobId}:
    get:
      tags:
        - Ingestion
      summary: Get ingestion job status
      description: |
        Poll ingestion job status and progress.
        Job status updates in real-time during processing.

      operationId: getIngestionJobStatus
      parameters:
        - name: jobId
          in: path
          required: true
          description: Job identifier from upload response
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440005"
      responses:
        '200':
          description: Job status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatusResponse'
              examples:
                processing:
                  summary: Job in progress
                  value:
                    job_id: "550e8400-e29b-41d4-a716-446655440005"
                    command_id: "550e8400-e29b-41d4-a716-446655440008"
                    status: "processing"
                    progress_percentage: 45.0
                    current_stage: "Generating embeddings"
                    chunks_processed: 45
                    total_chunks: 100
                    started_at: "2025-01-14T12:00:00Z"
                    estimated_completion_at: "2025-01-14T12:02:00Z"
                completed:
                  summary: Job completed successfully
                  value:
                    job_id: "550e8400-e29b-41d4-a716-446655440005"
                    command_id: "550e8400-e29b-41d4-a716-446655440008"
                    status: "completed"
                    progress_percentage: 100.0
                    current_stage: "Completed"
                    chunks_processed: 100
                    total_chunks: 100
                    started_at: "2025-01-14T12:00:00Z"
                    completed_at: "2025-01-14T12:02:00Z"
                failed:
                  summary: Job failed with error
                  value:
                    job_id: "550e8400-e29b-41d4-a716-446655440005"
                    command_id: "550e8400-e29b-41d4-a716-446655440008"
                    status: "failed"
                    progress_percentage: 30.0
                    current_stage: "Failed during parsing"
                    chunks_processed: 0
                    total_chunks: 0
                    error_message: "Failed to parse PDF: corrupted file"
                    started_at: "2025-01-14T12:00:00Z"
                    completed_at: "2025-01-14T12:00:15Z"
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: "Job not found"
                correlationId: "550e8400-e29b-41d4-a716-446655440009"
                timestamp: "2025-01-14T12:00:00Z"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    DocumentUploadRequest:
      type: object
      required:
        - file
        - taxonomy_path
      properties:
        file:
          type: string
          format: binary
          description: Document file to upload
        taxonomy_path:
          type: array
          description: Hierarchical taxonomy path for classification
          items:
            type: string
            minLength: 1
            maxLength: 100
          minItems: 1
          maxItems: 10
          example: ["Technology", "AI", "Machine Learning"]
        source_url:
          type: string
          format: uri
          description: Original document source URL
          maxLength: 2000
          nullable: true
          example: "https://example.com/document.pdf"
        author:
          type: string
          description: Document author
          maxLength: 200
          nullable: true
          example: "John Doe"
        language:
          type: string
          description: Document language code (ISO 639-1)
          pattern: '^[a-z]{2}$'
          default: "ko"
          example: "en"
        priority:
          type: integer
          description: Processing priority (1=highest, 10=lowest)
          minimum: 1
          maximum: 10
          default: 5
          example: 5
    IngestionJobResponse:
      type: object
      required:
        - job_id
        - status
        - message
      properties:
        job_id:
          type: string
          format: uuid
          description: Unique job identifier
          example: "550e8400-e29b-41d4-a716-446655440005"
        status:
          type: string
          enum: [pending, processing, completed, failed]
          description: Current job status
          example: "pending"
        estimated_completion_minutes:
          type: integer
          description: Estimated completion time in minutes
          minimum: 0
          example: 2
        message:
          type: string
          description: Human-readable status message
          maxLength: 500
          example: "Document accepted for processing"
    JobStatusResponse:
      type: object
      required:
        - job_id
        - command_id
        - status
        - progress_percentage
      properties:
        job_id:
          type: string
          format: uuid
          description: Unique job identifier
          example: "550e8400-e29b-41d4-a716-446655440005"
        command_id:
          type: string
          format: uuid
          description: Internal command identifier
          example: "550e8400-e29b-41d4-a716-446655440008"
        status:
          type: string
          enum: [pending, processing, completed, failed]
          description: Current job status
          example: "processing"
        progress_percentage:
          type: number
          format: float
          description: Job completion percentage
          minimum: 0.0
          maximum: 100.0
          example: 45.0
        current_stage:
          type: string
          description: Current processing stage
          maxLength: 100
          nullable: true
          example: "Generating embeddings"
        chunks_processed:
          type: integer
          description: Number of chunks processed
          minimum: 0
          example: 45
        total_chunks:
          type: integer
          description: Total number of chunks
          minimum: 0
          example: 100
        error_message:
          type: string
          description: Error message if failed
          maxLength: 1000
          nullable: true
        started_at:
          type: string
          format: date-time
          description: Job start timestamp
          nullable: true
          example: "2025-01-14T12:00:00Z"
        completed_at:
          type: string
          format: date-time
          description: Job completion timestamp
          nullable: true
          example: "2025-01-14T12:02:00Z"
        estimated_completion_at:
          type: string
          format: date-time
          description: Estimated completion timestamp
          nullable: true
          example: "2025-01-14T12:02:00Z"
    ErrorResponse:
      type: object
      required:
        - detail
        - correlationId
        - timestamp
      properties:
        detail:
          type: string
          description: Error message
          minLength: 1
          maxLength: 1000
          example: "Search query cannot be empty"
        correlationId:
          type: string
          description: Correlation ID for error tracing
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        timestamp:
          type: string
          description: Error timestamp
          format: date-time
          example: "2025-01-14T12:00:00Z"
        code:
          type: string
          description: Error code
          minLength: 1
          maxLength: 100
          example: "INVALID_REQUEST"
          nullable: true
        details:
          type: object
          description: Additional error details
          additionalProperties: true
          nullable: true
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication

    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
security:
  - ApiKeyAuth: []
  - BearerAuth: []